+++ lib/miq_automation_engine/engine/miq_ae_method_service/miq_ae_service_front.rb
@@ -1,0 +1,13 @@
+module MiqAeMethodService
+  class MiqAeServiceFront
+    include DRbUndumped
+    attr_accessor :workspace
+    def initialize(workspace)
+      @workspace = workspace
+    end
+
+    def find(id)
+      MiqAeService.find(id)
+    end
+  end
+end
