+++ lib/miq_automation_engine/engine/miq_ae_service.rb
@@ -3,18 +3,6 @@ require_relative 'miq_ae_service/miq_ae_service_object_common'
 require_relative 'miq_ae_service/miq_ae_service_vmdb'
 require_relative 'miq_ae_service/miq_ae_service_rbac'
 module MiqAeMethodService
-  class MiqAeServiceFront
-    include DRbUndumped
-    attr_accessor :workspace
-    def initialize(workspace)
-      @workspace = workspace
-    end
-
-    def find(id)
-      MiqAeService.find(id)
-    end
-  end
-
   class MiqAeService
     include Vmdb::Logging
     include DRbUndumped
